+++ core-services/account-service/src/main/java/com/piggymetrics/account/AccountApplication.java
@@ -1,32 +1,20 @@
 package com.piggymetrics.account;
 
-import com.piggymetrics.account.service.security.CustomUserInfoTokenServices;
 import feign.RequestInterceptor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.security.oauth2.resource.ResourceServerProperties;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.cloud.security.oauth2.client.feign.OAuth2FeignRequestInterceptor;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Primary;
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.oauth2.client.DefaultOAuth2ClientContext;
 import org.springframework.security.oauth2.client.OAuth2RestTemplate;
 import org.springframework.security.oauth2.client.token.grant.client.ClientCredentialsResourceDetails;
 import org.springframework.security.oauth2.config.annotation.web.configuration.EnableOAuth2Client;
 import org.springframework.security.oauth2.config.annotation.web.configuration.EnableResourceServer;
-import org.springframework.security.oauth2.config.annotation.web.configuration.ResourceServerConfigurerAdapter;
-import org.springframework.security.oauth2.config.annotation.web.configurers.ResourceServerSecurityConfigurer;
-import org.springframework.security.oauth2.provider.token.DefaultTokenServices;
-import org.springframework.security.oauth2.provider.token.ResourceServerTokenServices;
-import org.springframework.security.oauth2.provider.token.TokenStore;
-import org.springframework.security.oauth2.provider.token.store.JwtAccessTokenConverter;
-import org.springframework.security.oauth2.provider.token.store.JwtTokenStore;
 
 @SpringBootApplication
 @EnableResourceServer
@@ -35,9 +23,7 @@ import org.springframework.security.oauth2.provider.token.store.JwtTokenStore;
 @EnableFeignClients
 @EnableGlobalMethodSecurity(prePostEnabled = true)
 @EnableConfigurationProperties
-public class AccountApplication extends ResourceServerConfigurerAdapter {
-
-  @Autowired private ResourceServerProperties sso;
+public class AccountApplication {
 
   public static void main(String[] args) {
     SpringApplication.run(AccountApplication.class, args);
@@ -58,46 +44,5 @@ public class AccountApplication extends ResourceServerConfigurerAdapter {
   @Bean
   public OAuth2RestTemplate clientCredentialsRestTemplate() {
     return new OAuth2RestTemplate(clientCredentialsResourceDetails());
-  }
-
-//  @Bean
-//  public ResourceServerTokenServices tokenServices() {
-//    return new CustomUserInfoTokenServices(sso.getUserInfoUri(), sso.getClientId());
-//  }
-
-
-
-  @Override
-  public void configure(HttpSecurity http) throws Exception {
-    http.authorizeRequests()
-        .antMatchers("/", "/demo", "/actuator/**")
-        .permitAll()
-        .anyRequest()
-        .authenticated();
-  }
-
-  @Bean
-  @Primary
-  public DefaultTokenServices tokenServices() {
-    DefaultTokenServices defaultTokenServices = new DefaultTokenServices();
-    defaultTokenServices.setTokenStore(tokenStore());
-    return defaultTokenServices;
-  }
-
-  @Override
-  public void configure(ResourceServerSecurityConfigurer config) {
-    config.tokenServices(tokenServices());
-  }
-
-  @Bean
-  public TokenStore tokenStore() {
-    return new JwtTokenStore(accessTokenConverter());
-  }
-
-  @Bean
-  public JwtAccessTokenConverter accessTokenConverter() {
-    JwtAccessTokenConverter converter = new JwtAccessTokenConverter();
-    converter.setSigningKey("123456");
-    return converter;
   }
 }
