+++ core-services/account-service/src/main/java/com/piggymetrics/account/ResourceServerConfig.java
@@ -1,0 +1,32 @@
+package com.piggymetrics.account;
+
+import com.piggymetrics.account.service.security.CustomUserInfoTokenServices;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.security.oauth2.resource.ResourceServerProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.oauth2.config.annotation.web.configuration.EnableResourceServer;
+import org.springframework.security.oauth2.config.annotation.web.configuration.ResourceServerConfigurerAdapter;
+import org.springframework.security.oauth2.provider.token.ResourceServerTokenServices;
+
+@Configuration
+@EnableResourceServer
+public class ResourceServerConfig extends ResourceServerConfigurerAdapter {
+
+  @Autowired private ResourceServerProperties sso;
+
+  @Override
+  public void configure(HttpSecurity http) throws Exception {
+    http.authorizeRequests()
+        .antMatchers("/", "/demo", "/actuator/**")
+        .permitAll()
+        .anyRequest()
+        .authenticated();
+  }
+
+  @Bean
+  public ResourceServerTokenServices tokenServices() {
+    return new CustomUserInfoTokenServices(sso.getUserInfoUri(), sso.getClientId());
+  }
+}
